+++ components/LegislatorProfile/LegislatorSidebar.tsx
@@ -1,0 +1,14 @@
+import { Biography } from "./SidebarComponents/Biography"
+import { OtherTestimony } from "./SidebarComponents/OtherTestimony"
+import { UpcomingHearings } from "./SidebarComponents/UpcomingHearings"
+
+export function LegislatorSidebar() {
+  return (
+    <>
+      Sidebar Components
+      <OtherTestimony />
+      <UpcomingHearings />
+      <Biography />
+    </>
+  )
+}
