--- components/legislator/SidebarComponents/LegislatorSidebar.tsx
@@ -1,14 +1,0 @@
-import { Biography } from "./Biography"
-import { OtherTestimony } from "./OtherTestimony"
-import { UpcomingHearings } from "./UpcomingHearings"
-
-export function LegislatorSidebar() {
-  return (
-    <>
-      Sidebar Components
-      <OtherTestimony />
-      <UpcomingHearings />
-      <Biography />
-    </>
-  )
-}
