--- components/legislator/TabComponents/Bills.tsx
@@ -1,3 +1,0 @@
-export function Bills() {
-  return <div>- Bills</div>
-}
