--- components/legislator/TabComponents/District.tsx
@@ -1,3 +1,0 @@
-export function District() {
-  return <div>- District</div>
-}
