--- components/legislator/TabComponents/Elections.tsx
@@ -1,3 +1,0 @@
-export function Elections() {
-  return <div>- Elections</div>
-}
