--- components/legislator/TabComponents/Finance.tsx
@@ -1,3 +1,0 @@
-export function Finance() {
-  return <div>- Finance</div>
-}
