--- components/legislator/TabComponents/Priorities.tsx
@@ -1,3 +1,0 @@
-export function Priorities() {
-  return <div>- Priorities</div>
-}
