--- components/legislator/TabComponents/Votes.tsx
@@ -1,3 +1,0 @@
-export function Votes() {
-  return <div>- Votes</div>
-}
